Refactor AuthService to invert JWT verification dependency
Extracted JWT decoding logic from `AuthService` into a new infrastructure class `SupabaseJWTVerifier` that implements `TokenVerifierProtocol`. This adheres to Clean Architecture by removing infrastructure details (like the PyJWT library and HTTP requests for JWKS) from the application service.

- Created `TokenVerifierProtocol` in `app.domain.auth.interfaces`
- Implemented `SupabaseJWTVerifier` in `app.infrastructure.auth.jwt_verifier`
- Updated `AuthService` to use constructor injection for the verifier
- Updated FastAPI dependencies and websocket endpoint to use the verifier singleton
- Updated test suites to reflect the new architecture
