+++ backend/app/domain/auth/interfaces.py
@@ -6,6 +6,15 @@ from typing import Protocol
 from uuid import UUID
 
 from app.domain.auth.entities import Passkey, PasskeyCreate
+from app.domain.auth.schemas import JWTPayload
+
+
+class TokenVerifierProtocol(Protocol):
+    """Protocol for JWT token verification."""
+
+    async def verify_token(self, token: str) -> JWTPayload:
+        """Verify the token and return the payload."""
+        pass
 
 
 class AuthRepositoryProtocol(Protocol):
